backend: move errors to errors.go
